backend/requirements.txt
fastapi
uvicorn
pydantic
google-cloud-storage
google-generativeai
pydantic-settings
